move preview-related code to PreviewStyling.kt
